+++ src/main/java/com/board_of_ads/model/posting/autoTransport/partsAndAccessories/PartAndAccessorie.java
@@ -1,10 +1,12 @@
-package com.board_of_ads.model.posting.autoTransport;
+package com.board_of_ads.model.posting.autoTransport.partsAndAccessories;
 
+import com.board_of_ads.model.posting.autoTransport.AutoTransport;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Inheritance;
 import javax.persistence.InheritanceType;
@@ -19,6 +21,7 @@ import javax.persistence.Table;
 @Table(name = "part_and_accessories")
 public class PartAndAccessorie extends AutoTransport {
 
+    @Column
     private String typeOfPosting;
 
 }
