+++ src/main/java/com/board_of_ads/model/posting/autoTransport/partsAndAccessories/tiresDisksWheels/Tires.java
@@ -1,10 +1,12 @@
 package com.board_of_ads.model.posting.autoTransport.partsAndAccessories.tiresDisksWheels;
 
+import com.board_of_ads.model.posting.autoTransport.partsAndAccessories.PartAndAccessorie;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
@@ -14,13 +16,20 @@ import javax.persistence.Table;
 @AllArgsConstructor
 @Entity
 @Table(name = "tires")
-public class Tires extends Caps {
+public class Tires extends PartAndAccessorie {
 
+    @Column
+    private Byte diameter;
+
+    @Column
     private String manufacturer;
 
+    @Column
     private Short profileWidth;
 
+    @Column
     private String season;
 
+    @Column
     private Short profileHeight;
 }
