+++ src/main/java/com/board_of_ads/model/posting/autoTransport/waterTransport/WaterTransport.java
@@ -1,5 +1,6 @@
-package com.board_of_ads.model.posting.autoTransport;
+package com.board_of_ads.model.posting.autoTransport.waterTransport;
 
+import com.board_of_ads.model.posting.autoTransport.AutoTransport;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
